gather parsing activities together
