+++ lidy-js/nodes/rulenode.js
@@ -1,12 +1,15 @@
 import { LidyNode } from "./lidynode.js"
-import { parse_rule } from "../parser/parse.js"
+import { isScalar, isMap, isSeq } from "yaml"
 
 export class RuleNode extends LidyNode {
-    constructor(ctx, rule_name, rule, current) {
+    constructor(ctx, rule_name, current, value) {
       super(ctx, rule_name, current)
-      this.childs.push(parse_rule(ctx, null, rule, current))
-      if (['string'].includes(rule_name)) {
-        ctx.syntaxError(current, `'${rule_name}' is not allowed as rule_name in Lidy Grammar (reserved keyword)`)
-      }
-    }
+      this.childs.push(value)
+      this.value = value
   }
+
+  static checkCurrent(current) {
+    return isScalar(current) || isMap(current) || isSeq(current)
+  }
+  
+}
