+++ lidy-js/nodes/scalars/binarynode.js
@@ -1,11 +1,11 @@
-import { ScalarNode } from "../scalarnode.js"
+import { ScalarNode } from "./scalarnode.js"
 import { isScalar  } from 'yaml'
 
 // BinaryNode manages values stored in base64 format
 export class BinaryNode extends ScalarNode {
   constructor(ctx, current) {
     super(ctx, 'binary', current)
-    if (checkCurrent(current)) {
+    if (BinaryNode.checkCurrent(current)) {
         this.value = current.value
     } else {
       throw ctx.syntaxError(current, `Error: value '${current ? current.value : ""}' is not a base64 string`)
@@ -17,7 +17,7 @@ export class BinaryNode extends ScalarNode {
   }
 
   static parse(ctx, current) {
-    if (checkCurrent(current)) { return new Base64Node(ctx, current) }
+    if (BinaryNode.checkCurrent(current)) { return new Base64Node(ctx, current) }
     return null
   }
