+++ lidy-js/nodes/scalars/booleannode.js
@@ -1,11 +1,11 @@
-import { ScalarNode } from "../scalarnode.js"
+import { ScalarNode } from "./scalarnode.js"
 import { isScalar  } from 'yaml'
 
 export class BooleanNode extends ScalarNode {
   constructor(ctx, current) {
     super(ctx, 'boolean', current)
     this.value = null
-    if (checkCurrent(current)) {
+    if (BooleanNode.checkCurrent(current)) {
           this.value = current.value
     } else {
       throw ctx.syntaxError(current, `Error: value '${current.value}' is not a boolean`)
@@ -13,11 +13,11 @@ export class BooleanNode extends ScalarNode {
   }
 
   static checkCurrent(current) {
-    return isScalar(current) && (typeof(current) == 'boolean')
+    return isScalar(current) && (typeof(current.value) == 'boolean')
   }
 
   static parse(ctx, current) {
-    if (checkCurrent(current)) { return new BooleanNode(ctx, current) }
+    if (BooleanNode.checkCurrent(current)) { return new BooleanNode(ctx, current) }
     return null
 
   }
