+++ lidy-js/nodes/scalars/floatnode.js
@@ -1,10 +1,10 @@
-import { ScalarNode } from "../scalarnode.js"
+import { ScalarNode } from "./scalarnode.js"
 import { isScalar  } from 'yaml'
 
 export class FloatNode extends ScalarNode {
   constructor(ctx, current) {
     super(ctx, 'float', current)
-    if (currentCheck(current)) {
+    if (FloatNode.checkCurrent(current)) {
       this.value = current.value
     } else {
       throw ctx.syntaxError(current, `Error: value '${current ? current.value : ""}' is not a number`)
@@ -16,7 +16,7 @@ export class FloatNode extends ScalarNode {
   }
 
   static parse(ctx, current) {
-    if (checkCurrent(current)) { return new FloatNode(ctx, current) }
+    if (FloatNode.checkCurrent(current)) { return new FloatNode(ctx, current) }
     return null
 
   }
