+++ lidy-js/nodes/scalars/timestampnode.js
@@ -1,17 +1,17 @@
-import { ScalarNode } from "../scalarnode.js"
+import { ScalarNode } from "./scalarnode.js"
 import { isScalar  } from 'yaml'
 
 export class TimestampNode extends ScalarNode {
   constructor(ctx, current, date) {
     super(ctx, 'timestamp', current)
-    if (this.checkCurrent(current)) {
+    if (TimestampNode.checkCurrent(current)) {
       this.value = new Date(current.value) // lidy-js accepts as timestamp same date format as javascript (simplified ISO8601 format) 
     } else {
       throw ctx.syntaxError(current, `Error: value '${(current) ? current.value : ""}' is not a timestamp in ISO9601 format`)
     }
   }
 
-  checkCurrent(current) {
+  static checkCurrent(current) {
     let iso8601regex=/\d{4}-\d{2}-\d{2}|\d{4}-\d{2}?-\d{2}?([Tt]|[ \t]+)\d{2}?:\d{2}:\d{2}(\.\d*)?(([ \t]*)Z|[-+]\d\d?(:\d{2})?)?/
     let ret = isScalar(current) && (typeof(current.value) == 'string') && iso8601regex.test(current.value)
     if (ret == true) {
@@ -25,7 +25,7 @@ export class TimestampNode extends ScalarNode {
   }
 
   static parse(ctx, current) {
-    if (checkCurrent(current)) { return new TimestampNode(ctx, current) }
+    if (TimestampNode.checkCurrent(current)) { return new TimestampNode(ctx, current) }
     return null
 
   }
