+++ lidy-js/tests/jasmine.js
@@ -1,7 +1,7 @@
 import glob from 'glob';
 import Jasmine from 'jasmine';
 
-glob('./**/*.js', async (err, files) => {
+glob('./**/*-spec.js', async (err, files) => {
   if (err) { throw err; }
   let jasmine = new Jasmine();
   await Promise.all(files.map(f => import(f)));
