+++ lidy-js/tests/scalars/anynode-spec.js
@@ -1,4 +1,4 @@
-import { parse } from '../../parse.js'
+import { parse } from '../../parser/parse.js'
 
 describe("Lidy scalars ->", function() {
 
@@ -11,19 +11,19 @@ describe("Lidy scalars ->", function() {
             function() { expect( parse({src_data: "12.5", dsl_data: "main: any"}).contents.getChild(0).value).toEqual(12.5)})
 
         it("A boolean value",
-            function() { expect( parse({src_data: "True", dsl_data: "main: any"}).contents.getChild(0).value).toEqual(true)})
+            function() { expect( parse({src_data: "true", dsl_data: "main: any"}).contents.getChild(0).value).toEqual(true)})
 
         it("A map value",
-            function() { expect( parse({src_data: "{ un: 1, deux: 2 }", dsl_data: "main: any"}).contents.getChild(0).value).toEqual({un: 1, deux:2})})
+            function() { expect( parse({src_data: "{ un: 1, deux: 2 }", dsl_data: "main: any"}).contents.getChild(0).value["un"].value).toEqual(1)})
 
         it("A list value",
-            function() { expect( parse({src_data: '[ un, 1, deux, null ]', dsl_data: "main: any"}).contents.getChild(0).value).toEqual([ "un", 1, "deux", null ])})
+            function() { expect( parse({src_data: '[ un, 1, deux, null, 2 ]', dsl_data: "main: any"}).contents.getChild(0).value[4].value).toEqual(2)})
 
         it("A null value",
             function() { expect( parse({src_data: '~', dsl_data: "main: any"}).contents.getChild(0).value).toEqual(null)})
 
         it("complex type",
-            function() { expect( parse({src_data: "{ un: [1, { un: '1' } ], deux: 2 }", dsl_data: "main: any"}).contents.getChild(0).value).toEqual({un: [1, { un: '1'}], deux:2})})
+            function() { expect( parse({src_data: "{ un: [1, { un: '1' } ], deux: 2 }", dsl_data: "main: any"}).contents.getChild(0).value["un"].value[1].value["un"].value).toEqual('1')})
 
         it("values that could be strings or something else (timestamp or base64) are parsed as strings by 'any'",
             function() { expect( parse({src_data: "2021-01-12", dsl_data: "main: any"}).contents.getChild(0).value).toEqual("2021-01-12")})
